+++ CHANGELOG.md
@@ -4,7 +4,10 @@ Changelog
 HEAD
 ----------------
 
-* logging: updated logging module to use structlog's threadlocal implementation
+* bumped dependencies
+* updated logging module to use structlog's threadlocal implementation
+* added the ability to customize the log format, with a bunch of variables
+  that can be used.
 
 0.9.7 13-03-2021
 ----------------